restructure: because my debugger is throwing a hissyfit. Also add some additional functionality.
