+++ bots/example_bots/grid_types/grid_params.py
@@ -1,4 +1,5 @@
 from _decimal import Decimal
+from typing import Optional
 
 from rubi import OrderSide
 
@@ -13,7 +14,10 @@ class GridParams:
         grid_spread_in_quote: Decimal,
         level_spread_multiplier: Decimal,
         number_levels: int,
-        level_allocation_multiplier: Decimal
+        base_level_size: Decimal,
+        # order related
+        min_order_size_in_quote: Decimal,
+        allowed_order_drift: Optional[Decimal] = None
     ):
         self.base_asset_amount = starting_base_asset_amount
         self.quote_asset_amount = starting_quote_asset_amount
@@ -21,17 +25,24 @@ class GridParams:
         self.grid_spread_in_quote = grid_spread_in_quote
         self.level_spread_multiplier = level_spread_multiplier
         self.number_levels = number_levels
-        self.level_allocation_multiplier = level_allocation_multiplier
+        self.base_level_size = base_level_size
 
-    def base_level_size(self, price: Decimal) -> Decimal:
-        total_grid_amount = self.base_asset_amount + self.quote_asset_amount / price
+        self.scale_factor = self._calculate_scale_factor()
 
-        base_level_multiplier = self.number_levels + (1 * self.level_allocation_multiplier) ** self.number_levels
+        # transaction related
+        self.min_order_size_in_quote = min_order_size_in_quote
+        self.allowed_order_drift = allowed_order_drift
 
-        return (total_grid_amount / 2) / base_level_multiplier
+    def get_level_size(self, level: int) -> Decimal:
+        return self.base_level_size + level * self.base_level_size * (1 + self.scale_factor)
 
-    def get_level_size(self, level: int, price: Decimal) -> Decimal:
-        return self.base_level_size(price=price) * (1 + self.level_allocation_multiplier) ** level
+    def _calculate_scale_factor(self):
+        total_grid_size = self.base_asset_amount + self.quote_asset_amount / self.mid_price
+
+        scale_factor = ((((total_grid_size / 2) / self.base_level_size) - self.number_levels) / (
+            Decimal((self.number_levels / 2) * (self.number_levels - 1)))) - 1
+
+        return scale_factor
 
     def get_level_price(self, side: OrderSide, level: int) -> Decimal:
         return self.mid_price - side.sign() * (self.grid_spread_in_quote / 2) * (level + 1) * (
