+++ bots/example_bots/helpers/inventory_manager.py
